requirements.txt
# Core dependencies
pandas==2.0.3
mysql-connector-python==8.3.0
python-docx==1.1.0
docx2txt==0.8
python-dotenv==1.0.1

# SQLAlchemy (optional, for future migration)
sqlalchemy==2.0.23
pymysql==1.1.0

# Development dependencies
black==23.12.1
flake8==7.0.0
pytest==7.4.4